+++ s3_lifecycle.tf
@@ -1,0 +1,122 @@
+# 3rd party software is contained in this file. Please see NOTICE file.
+# Minimally modified to fit the needs of this product
+
+# Variables with type `any` should be jsonencode()'d when value is coming from Terragrunt
+locals {
+  lifecycle_rules = try(jsondecode(var.lifecycle_rule), var.lifecycle_rule)
+}
+
+resource "aws_s3_bucket_lifecycle_configuration" "this" {
+  count = length(local.lifecycle_rules) > 0 ? 1 : 0
+
+  bucket                = aws_s3_bucket.kinesis_firehose_s3_bucket.id
+  expected_bucket_owner = var.expected_bucket_owner
+
+  dynamic "rule" {
+    for_each = local.lifecycle_rules
+
+    content {
+      id     = try(rule.value.id, null)
+      status = try(rule.value.enabled ? "Enabled" : "Disabled", tobool(rule.value.status) ? "Enabled" : "Disabled", title(lower(rule.value.status)))
+
+      # Max 1 block - abort_incomplete_multipart_upload
+      dynamic "abort_incomplete_multipart_upload" {
+        for_each = try([rule.value.abort_incomplete_multipart_upload_days], [])
+
+        content {
+          days_after_initiation = try(rule.value.abort_incomplete_multipart_upload_days, null)
+        }
+      }
+
+
+      # Max 1 block - expiration
+      dynamic "expiration" {
+        for_each = try(flatten([rule.value.expiration]), [])
+
+        content {
+          date                         = try(expiration.value.date, null)
+          days                         = try(expiration.value.days, null)
+          expired_object_delete_marker = try(expiration.value.expired_object_delete_marker, null)
+        }
+      }
+
+      # Several blocks - transition
+      dynamic "transition" {
+        for_each = try(flatten([rule.value.transition]), [])
+
+        content {
+          date          = try(transition.value.date, null)
+          days          = try(transition.value.days, null)
+          storage_class = transition.value.storage_class
+        }
+      }
+
+      # Max 1 block - noncurrent_version_expiration
+      dynamic "noncurrent_version_expiration" {
+        for_each = try(flatten([rule.value.noncurrent_version_expiration]), [])
+
+        content {
+          newer_noncurrent_versions = try(noncurrent_version_expiration.value.newer_noncurrent_versions, null)
+          noncurrent_days           = try(noncurrent_version_expiration.value.days, noncurrent_version_expiration.value.noncurrent_days, null)
+        }
+      }
+
+      # Several blocks - noncurrent_version_transition
+      dynamic "noncurrent_version_transition" {
+        for_each = try(flatten([rule.value.noncurrent_version_transition]), [])
+
+        content {
+          newer_noncurrent_versions = try(noncurrent_version_transition.value.newer_noncurrent_versions, null)
+          noncurrent_days           = try(noncurrent_version_transition.value.days, noncurrent_version_transition.value.noncurrent_days, null)
+          storage_class             = noncurrent_version_transition.value.storage_class
+        }
+      }
+
+      # Max 1 block - filter - without any key arguments or tags
+      dynamic "filter" {
+        for_each = length(try(flatten([rule.value.filter]), [])) == 0 ? [true] : []
+
+        content {
+          #          prefix = ""
+        }
+      }
+
+      # Max 1 block - filter - with one key argument or a single tag
+      dynamic "filter" {
+        for_each = [for v in try(flatten([rule.value.filter]), []) : v if max(length(keys(v)), length(try(rule.value.filter.tags, rule.value.filter.tag, []))) == 1]
+
+        content {
+          object_size_greater_than = try(filter.value.object_size_greater_than, null)
+          object_size_less_than    = try(filter.value.object_size_less_than, null)
+          prefix                   = try(filter.value.prefix, null)
+
+          dynamic "tag" {
+            for_each = try(filter.value.tags, filter.value.tag, [])
+
+            content {
+              key   = tag.key
+              value = tag.value
+            }
+          }
+        }
+      }
+
+      # Max 1 block - filter - with more than one key arguments or multiple tags
+      dynamic "filter" {
+        for_each = [for v in try(flatten([rule.value.filter]), []) : v if max(length(keys(v)), length(try(rule.value.filter.tags, rule.value.filter.tag, []))) > 1]
+
+        content {
+          and {
+            object_size_greater_than = try(filter.value.object_size_greater_than, null)
+            object_size_less_than    = try(filter.value.object_size_less_than, null)
+            prefix                   = try(filter.value.prefix, null)
+            tags                     = try(filter.value.tags, filter.value.tag, null)
+          }
+        }
+      }
+    }
+  }
+
+  # Must have bucket versioning enabled first
+  depends_on = [aws_s3_bucket_versioning.kinesis_firehose_s3_bucket_versioning]
+}